+++ TerminalCommander/Patches/TerminalHotkeys.cs
@@ -88,6 +88,11 @@ namespace TerminalCommander.Patches
                     {
                         SetTerminalText(t, TerminalCommands.InverseTeleportCommand());
                     }
+                    //Emergency Teleport Hot Key
+                    else if (BepInEx.UnityInput.Current.GetKeyDown(commanderSource.Configs.EmergencyTeleportKey))
+                    {
+                        SetTerminalText(t, TerminalCommands.EmergencyTeleportCommand());
+                    }
                 }
             }
             catch(Exception ex)
@@ -102,7 +107,7 @@ namespace TerminalCommander.Patches
             //TerminalNode tn = t.terminalNodes.specialNodes[20];
             //StartOfRound.Instance.mapScreen.SwitchRadarTargetForward(callRPC: true);
             //t.LoadNewNode(tn);
-
+         
             string cmd = "switch";
 
             t.screenText.text += cmd;
@@ -120,12 +125,14 @@ namespace TerminalCommander.Patches
             if(!commanderSource.Configs.AllowBigDoors)
             {
                 SetTerminalText(t, "This command has been disabled by the company.\n\n");
+                t.terminalAudio.PlayOneShot(commanderSource.Audio.errorAudio);
                 return;
             }
             if(d < commanderSource.LastDoorEvent)
             {
                 var ts = commanderSource.LastDoorEvent - d;
                 SetTerminalText(t, $"Door signal cool down time remaining: {Math.Round(ts.TotalSeconds)} seconds.\n\n");
+                t.terminalAudio.PlayOneShot(commanderSource.Audio.errorAudio);
                 return;
             }
             TerminalAccessibleObject[] taos = (from x in UnityEngine.Object.FindObjectsOfType<TerminalAccessibleObject>()
@@ -167,12 +174,15 @@ namespace TerminalCommander.Patches
             if (!commanderSource.Configs.AllowJamming)
             {
                 SetTerminalText(t, "This command has been disabled by the company.\n\n");
+                t.terminalAudio.PlayOneShot(commanderSource.Audio.errorAudio);
                 return;
             }
             if (d < commanderSource.LastJamEvent)
             {
                 var ts = commanderSource.LastJamEvent - d;
                 SetTerminalText(t, $"Jammer cool down time remaining: {Math.Round(ts.TotalSeconds)} seconds.\n\n");
+                t.terminalAudio.PlayOneShot(commanderSource.Audio.errorAudio);
+             
                 return;
             }
             TerminalAccessibleObject[] taos = (from x in UnityEngine.Object.FindObjectsOfType<TerminalAccessibleObject>()
@@ -188,10 +198,10 @@ namespace TerminalCommander.Patches
             }
 
             SetTerminalText(t, "Jamming turrets and land mines\n\n");
-
+            t.terminalAudio.PlayOneShot(commanderSource.Audio.jammerAudio);
             t.terminalAudio.PlayOneShot(t.codeBroadcastSFX, 1f);
             t.codeBroadcastAnimator.SetTrigger("display");
-
+         
             commanderSource.LastJamEvent = DateTime.Now;
 
             logSource.LogInfo($"{Commander.modName} TerminalAccessibleObjects Called: Count{taos.Count()} - ({string.Join(", ", items)})");