+++ TerminalCommander/Plugin.cs
@@ -3,8 +3,9 @@ using BepInEx.Configuration;
 using BepInEx.Logging;
 using HarmonyLib;
 using TerminalCommander.Patches;
-using BepInEx.Bootstrap;
-using System.Security.Cryptography;
+using UnityEngine;
+using UnityEngine.Networking;
+
 using System;
 
 
@@ -15,7 +16,7 @@ namespace TerminalCommander
     {
         public const string modGUID = "Tronald.TerminalCommander";
         public const string modName = "TerminalCommander";
-        public const string modVersion = "1.8.0";
+        public const string modVersion = "1.9.0";
 
         private readonly Harmony harmony = new Harmony(modGUID);
 
@@ -27,8 +28,10 @@ namespace TerminalCommander
 
         public DateTime LastJamEvent = new DateTime();
         public DateTime LastDoorEvent = new DateTime();
+        public bool EmergencyTPUsed = false;
 
         internal ManualLogSource log;
+        public AudioManager Audio;
 
         void Awake()
         {
@@ -37,18 +40,21 @@ namespace TerminalCommander
             log = BepInEx.Logging.Logger.CreateLogSource(modGUID);
             log.LogInfo($"{modName} is loaded!");
 
-            Configs.Set_Configs(this);           
+            Configs.Set_Configs(this);
 
             TerminalHotkeys.SetSource(this);
             TerminalCommands.SetSource(this);
-            GameManagement.SetSource(this);
             RoundManagerPatch.SetSource(this);
+            ChatManagerPatch.SetSource(this);
+
+            Audio = new AudioManager();
+            Audio.LoadAudio();
 
             harmony.PatchAll(typeof(Commander));
             harmony.PatchAll(typeof(TerminalHotkeys));
             harmony.PatchAll(typeof(TerminalCommands));
-            harmony.PatchAll(typeof(GameManagement));
-            harmony.PatchAll(typeof (RoundManagerPatch));
-        }      
+            harmony.PatchAll(typeof(RoundManagerPatch));
+            harmony.PatchAll(typeof(ChatManagerPatch));
+        }
     }    
 }